Add files via upload
This React-based project offers a versatile text manipulation tool that allows users to convert text to uppercase or lowercase, clear text, reverse it, and more. Key features include a responsive alert system for notifications, a dynamic dark/light mode toggle, and a detailed about page with an interactive accordion for information. The app leverages Bootstrap for sleek, responsive design and React for efficient, state-driven functionality. The project is designed to enhance user experience by providing various text formatting options and an engaging interface.